requirements.txt
bottle==0.10.9
gevent==0.13.7
greenlet==0.3.4
igo-python==0.9.2
wsgiref==0.1.2
